fix: address review comments and add welcome step to walkthrough
Address review comments on PR #2712:
- Move WALKTHROUGH_TARGETS to side-effect-free constants/walkthrough.ts
  so pages don't pull driver.js via the hooks barrel
- Guard consoleAccount instead of non-null assertions in use-walkthrough
- Remove blockEnterSubmit listener in onDestroyed as a safety net
- Add missing stdlib imports to the Go SDK snippet

Also show a welcome dialog before the auto-started walkthrough for
first-time users.
