+++ ui/dashboard/src/@locales/ja/common.json
@@ -432,6 +432,11 @@
       "title": "キーを作成",
       "description": "フォームの入力が完了したら、送信ボタンをクリックしてAPIキーを作成しましょう。"
     },
-    "skip": "ウォークスルーをスキップ"
+    "skip": "ウォークスルーをスキップ",
+    "welcome": {
+      "title": "👋 Bucketeerへようこそ！",
+      "description": "Bucketeerはフィーチャーフラグと実験のプラットフォームです。短いツアーで最初のフィーチャーフラグをリリースしてみましょう。数分で完了します。",
+      "start": "ツアーを開始"
+    }
   }
 }
