+++ ui/dashboard/src/components/navigation/index.tsx
@@ -3,7 +3,8 @@ import logo from 'assets/logos/logo-white.svg';
 import { useAuth, getCurrentEnvironment } from 'auth';
 import { WALKTHROUGH_ENABLED } from 'configs';
 import * as ROUTING from 'constants/routing';
-import { useToggleOpen, useWalkthrough, WALKTHROUGH_TARGETS } from 'hooks';
+import { WALKTHROUGH_TARGETS } from 'constants/walkthrough';
+import { useToggleOpen, useWalkthrough } from 'hooks';
 import { useTranslation } from 'i18n';
 import compact from 'lodash/compact';
 import flatMapDeep from 'lodash/flatMapDeep';
