+++ ui/dashboard/src/pages/api-keys/api-key-modal/api-key-create-update-modal/index.tsx
@@ -3,7 +3,8 @@ import { FormProvider, SubmitHandler, useForm } from 'react-hook-form';
 import { apiKeyCreator, APIKeyResponse, apiKeyUpdater } from '@api/api-key';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { useAuthAccess } from 'auth';
-import { useToast, WALKTHROUGH_TARGETS } from 'hooks';
+import { WALKTHROUGH_TARGETS } from 'constants/walkthrough';
+import { useToast } from 'hooks';
 import useFormSchema, { FormSchemaProps } from 'hooks/use-form-schema';
 import useOptions from 'hooks/use-options';
 import { useUnsavedLeavePage } from 'hooks/use-unsaved-leave-page';
