+++ ui/dashboard/src/pages/api-keys/page-content.tsx
@@ -2,7 +2,8 @@ import { useCallback, useEffect, useMemo } from 'react';
 import { IconAddOutlined } from 'react-icons-material-design';
 import { useAuthAccess } from 'auth';
 import { DOCUMENTATION_LINKS } from 'constants/documentation-links';
-import { usePartialState, useToggleOpen, WALKTHROUGH_TARGETS } from 'hooks';
+import { WALKTHROUGH_TARGETS } from 'constants/walkthrough';
+import { usePartialState, useToggleOpen } from 'hooks';
 import { useTranslation } from 'i18n';
 import isNil from 'lodash/isNil';
 import pickBy from 'lodash/pickBy';
